Moving client telemetry utils to internal
